Simplify video capture
