+++ Sources/LiveKit/Extensions/RTCVideoCapturerDelegate+Buffer.swift
@@ -68,91 +68,6 @@ extension CGImagePropertyOrientation {
     }
 }
 
-extension LKRTCVideoCapturerDelegate {
-    typealias OnResolveSourceDimensions = (Dimensions) -> Void
-    typealias OnDidCreateFrame = (LKRTCVideoFrame) -> Void
-
-    /// capture a `CVPixelBuffer`, all other capture methods call this method internally.
-    func capturer(_ capturer: LKRTCVideoCapturer,
-                  didCapture pixelBuffer: CVPixelBuffer,
-                  timeStampNs: Int64 = VideoCapturer.createTimeStampNs(),
-                  rotation: RTCVideoRotation = ._0,
-
-                  onDidCreateFrame: OnDidCreateFrame? = nil)
-    {
-        // check if pixel format is supported by WebRTC
-        let pixelFormat = CVPixelBufferGetPixelFormatType(pixelBuffer)
-        guard VideoCapturer.supportedPixelFormats.contains(where: { $0.uint32Value == pixelFormat }) else {
-            // kCVPixelFormatType_420YpCbCr8BiPlanarVideoRange
-            // kCVPixelFormatType_420YpCbCr8BiPlanarFullRange
-            // kCVPixelFormatType_32BGRA
-            // kCVPixelFormatType_32ARGB
-            logger.log("Skipping capture for unsupported pixel format: \(pixelFormat.toString())", .warning,
-                       type: type(of: self))
-            return
-        }
-
-        let sourceDimensions = Dimensions(width: Int32(CVPixelBufferGetWidth(pixelBuffer)),
-                                          height: Int32(CVPixelBufferGetHeight(pixelBuffer)))
-
-        guard sourceDimensions.isEncodeSafe else {
-            logger.log("Skipping capture for dimensions: \(sourceDimensions)", .warning,
-                       type: type(of: self))
-            return
-        }
-
-        let rtcBuffer = LKRTCCVPixelBuffer(pixelBuffer: pixelBuffer)
-        let rtcFrame = LKRTCVideoFrame(buffer: rtcBuffer,
-                                       rotation: rotation,
-                                       timeStampNs: timeStampNs)
-
-        self.capturer(capturer, didCapture: rtcFrame)
-        onDidCreateFrame?(rtcFrame)
-    }
-
-    /// capture a `CMSampleBuffer`
-    func capturer(_ capturer: LKRTCVideoCapturer,
-                  didCapture sampleBuffer: CMSampleBuffer,
-
-                  onDidCreateFrame: OnDidCreateFrame? = nil)
-    {
-        // check if buffer is ready
-        guard CMSampleBufferGetNumSamples(sampleBuffer) == 1,
-              CMSampleBufferIsValid(sampleBuffer),
-              CMSampleBufferDataIsReady(sampleBuffer)
-        else {
-            logger.log("Failed to capture, buffer is not ready", .warning, type: type(of: self))
-            return
-        }
-
-        // attempt to determine rotation information if buffer is coming from ReplayKit
-        var rotation: RTCVideoRotation?
-        if #available(macOS 11.0, *) {
-            // Check rotation tags. Extensions see these tags, but `RPScreenRecorder` does not appear to set them.
-            // On iOS 12.0 and 13.0 rotation tags (other than up) are set by extensions.
-            if let sampleOrientation = CMGetAttachment(sampleBuffer, key: RPVideoSampleOrientationKey as CFString, attachmentModeOut: nil),
-               let coreSampleOrientation = sampleOrientation.uint32Value
-            {
-                rotation = CGImagePropertyOrientation(rawValue: coreSampleOrientation)?.toRTCRotation()
-            }
-        }
-
-        guard let pixelBuffer = CMSampleBufferGetImageBuffer(sampleBuffer) else {
-            logger.log("Failed to capture, pixel buffer not found", .warning, type: type(of: self))
-            return
-        }
-
-        let timeStamp = CMSampleBufferGetPresentationTimeStamp(sampleBuffer)
-        let timeStampNs = Int64(CMTimeGetSeconds(timeStamp) * Double(NSEC_PER_SEC))
-
-        self.capturer(capturer,
-                      didCapture: pixelBuffer,
-                      timeStampNs: timeStampNs,
-                      rotation: rotation ?? ._0,
-                      onDidCreateFrame: onDidCreateFrame)
-    }
-}
-
 extension CVPixelBuffer {
     func toDimensions() -> Dimensions {
         Dimensions(width: Int32(CVPixelBufferGetWidth(self)),
