+++ Sources/LiveKit/Track/Capturers/BufferCapturer.swift
@@ -45,12 +45,18 @@ public class BufferCapturer: VideoCapturer {
 
     /// Capture a ``CMSampleBuffer``.
     public func capture(_ sampleBuffer: CMSampleBuffer) {
-        capture(sampleBuffer: sampleBuffer, capturer: capturer, options: options)
+        capture(sampleBuffer: sampleBuffer,
+                capturer: capturer,
+                options: options)
     }
 
     /// Capture a ``CVPixelBuffer``.
     public func capture(_ pixelBuffer: CVPixelBuffer, timeStampNs: Int64 = VideoCapturer.createTimeStampNs(), rotation: VideoRotation = ._0) {
-        capture(pixelBuffer: pixelBuffer, capturer: capturer, timeStampNs: timeStampNs, rotation: rotation, options: options)
+        capture(pixelBuffer: pixelBuffer,
+                capturer: capturer,
+                timeStampNs: timeStampNs,
+                rotation: rotation,
+                options: options)
     }
 }
