+++ Sources/LiveKit/Track/Capturers/VideoCapturer.swift
@@ -22,6 +22,10 @@ internal import LiveKitWebRTC
 @_implementationOnly import LiveKitWebRTC
 #endif
 
+#if canImport(ReplayKit)
+import ReplayKit
+#endif
+
 protocol VideoCapturerProtocol {
     var capturer: LKRTCVideoCapturer { get }
 }
@@ -202,17 +206,10 @@ extension VideoCapturer {
                  device: AVCaptureDevice? = nil,
                  options: VideoCaptureOptions)
     {
-        _processFrame(frame, capturer: capturer, device: device, options: options)
-    }
-
-    // Capture a CMSampleBuffer
-    func capture(sampleBuffer: CMSampleBuffer,
-                 capturer: LKRTCVideoCapturer,
-                 options: VideoCaptureOptions)
-    {
-        delegate?.capturer(capturer, didCapture: sampleBuffer) { [weak self] frame in
-            self?._processFrame(frame, capturer: capturer, device: nil, options: options)
-        }
+        _process(frame: frame,
+                 capturer: capturer,
+                 device: device,
+                 options: options)
     }
 
     // Capture a CVPixelBuffer
@@ -222,16 +219,83 @@ extension VideoCapturer {
                  rotation: VideoRotation = ._0,
                  options: VideoCaptureOptions)
     {
-        delegate?.capturer(capturer, didCapture: pixelBuffer, timeStampNs: timeStampNs, rotation: rotation.toRTCType()) { [weak self] frame in
-            self?._processFrame(frame, capturer: capturer, device: nil, options: options)
+        // check if pixel format is supported by WebRTC
+        let pixelFormat = CVPixelBufferGetPixelFormatType(pixelBuffer)
+        guard VideoCapturer.supportedPixelFormats.contains(where: { $0.uint32Value == pixelFormat }) else {
+            // kCVPixelFormatType_420YpCbCr8BiPlanarVideoRange
+            // kCVPixelFormatType_420YpCbCr8BiPlanarFullRange
+            // kCVPixelFormatType_32BGRA
+            // kCVPixelFormatType_32ARGB
+            logger.log("Skipping capture for unsupported pixel format: \(pixelFormat.toString())", .warning,
+                       type: type(of: self))
+            return
         }
+
+        let sourceDimensions = Dimensions(width: Int32(CVPixelBufferGetWidth(pixelBuffer)),
+                                          height: Int32(CVPixelBufferGetHeight(pixelBuffer)))
+
+        guard sourceDimensions.isEncodeSafe else {
+            logger.log("Skipping capture for dimensions: \(sourceDimensions)", .warning,
+                       type: type(of: self))
+            return
+        }
+
+        let rtcBuffer = LKRTCCVPixelBuffer(pixelBuffer: pixelBuffer)
+        let rtcFrame = LKRTCVideoFrame(buffer: rtcBuffer,
+                                       rotation: rotation.toRTCType(),
+                                       timeStampNs: timeStampNs)
+
+        capture(frame: rtcFrame,
+                capturer: capturer,
+                options: options)
+    }
+
+    // Capture a CMSampleBuffer
+    func capture(sampleBuffer: CMSampleBuffer,
+                 capturer: LKRTCVideoCapturer,
+                 options: VideoCaptureOptions)
+    {
+        // Check if buffer is ready
+        guard CMSampleBufferGetNumSamples(sampleBuffer) == 1,
+              CMSampleBufferIsValid(sampleBuffer),
+              CMSampleBufferDataIsReady(sampleBuffer)
+        else {
+            logger.log("Failed to capture, buffer is not ready", .warning, type: type(of: self))
+            return
+        }
+
+        // attempt to determine rotation information if buffer is coming from ReplayKit
+        var rotation: RTCVideoRotation?
+        if #available(macOS 11.0, *) {
+            // Check rotation tags. Extensions see these tags, but `RPScreenRecorder` does not appear to set them.
+            // On iOS 12.0 and 13.0 rotation tags (other than up) are set by extensions.
+            if let sampleOrientation = CMGetAttachment(sampleBuffer, key: RPVideoSampleOrientationKey as CFString, attachmentModeOut: nil),
+               let coreSampleOrientation = sampleOrientation.uint32Value
+            {
+                rotation = CGImagePropertyOrientation(rawValue: coreSampleOrientation)?.toRTCRotation()
+            }
+        }
+
+        guard let pixelBuffer = CMSampleBufferGetImageBuffer(sampleBuffer) else {
+            logger.log("Failed to capture, pixel buffer not found", .warning, type: type(of: self))
+            return
+        }
+
+        let timeStamp = CMSampleBufferGetPresentationTimeStamp(sampleBuffer)
+        let timeStampNs = Int64(CMTimeGetSeconds(timeStamp) * Double(NSEC_PER_SEC))
+
+        capture(pixelBuffer: pixelBuffer,
+                capturer: capturer,
+                timeStampNs: timeStampNs,
+                rotation: rotation?.toLKType() ?? ._0,
+                options: options)
     }
 
     // Process the captured frame
-    private func _processFrame(_ frame: LKRTCVideoFrame,
-                               capturer: LKRTCVideoCapturer,
-                               device: AVCaptureDevice?,
-                               options: VideoCaptureOptions)
+    private func _process(frame: LKRTCVideoFrame,
+                          capturer: LKRTCVideoCapturer,
+                          device: AVCaptureDevice?,
+                          options: VideoCaptureOptions)
     {
         if _state.isFrameProcessingBusy {
             log("Frame processing hasn't completed yet, skipping frame...", .warning)
